Tweak title to include domain (but not "tot" version)
The domainGenerator generates the title for domains in the `title`
field of the return object of `data()`. As such, we should add
the logic to remove `tot` from the title there.

`/tot/` is generated from `tot.md` per Eleventy's mapping of
Markdown files to pages. For more information, see:
https://www.11ty.dev/docs/layouts/#front-matter-data-in-layouts

All subpages with the domains are generated by the `domainGenerator`.
For example, all subpages of `/tot/` are generated by `tot.11ty.js`.

Fixes #148